+++ src/main/kotlin/tools/ProgressBar.kt
@@ -10,9 +10,9 @@ class ProgressBar(private val maxValue: Int) {
 
     fun updateProgressBar() {
         currentValue += 1
-        val percentage = floor(currentValue.toFloat() / maxValue * 100).toInt()
+        val percentage = floor(currentValue.toDouble() / maxValue * 100).toInt()
         if (oldPercentage != percentage) {
-            val currentProgress = floor(currentValue.toFloat() / maxValue * 10).toInt()
+            val currentProgress = floor(currentValue.toDouble() / maxValue * 10).toInt()
             val missingProgress = barSize - currentProgress
             print("\r" + "|" + " - ".repeat(currentProgress) + "   ".repeat(missingProgress) + "|" + " $percentage%")
             oldPercentage = percentage
